mcp_server/staffing_backend/requirements.txt
supabase>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fastmcp>=0.9.0
PyPDF2>=3.0.1

